Add support to isolate brief description and discussion content.
The idea is provide a better stylesheet that will place @brief
content inside the description block, but separate the discussion
(longer description) below a title "description" the way Apple
does it. At the same time, we don't want to clutter the method
listing with @brief explanations so that feature has been turned
off.