Implement separation of @brief and main discussion text.
Similar to Apple documentation, when using the @brief tag the contents
are prepended to the description of each member function and then
followed by a detailed description with a new title "Discussion". The
brief content is NOT added to the method list at the top of the Class
page in order to keep the list compact.